requirements.txt
Flask==3.0.3
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.8.2
pytest==8.3.2
